requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
google-api-python-client==2.25.0
aiofiles==0.7.0
